refactor(adr-061): give reporter.py's appcompat projection its own owner
`reporter.py` sat two lines under the 2000-line hard cap. The one whole
concern inside it with no other caller is the application-compatibility
projection: `appcompat_to_json`, `appcompat_to_markdown`, and the five
`_appcompat_*_lines` section builders only they use. It answers a different
question from the rest of the module -- one application against one library
pair, rather than library against library -- and already has its own module
family (`appcompat.py`, `appcompat_html.py`).

Moved to `report/appcompat_report.py`, classified `report` by virtue of
living in the package. 1998 -> 1747 lines, and the `debt.yaml` baseline
drops with it.

The re-export is a lazy module-level `__getattr__`, not a static
`from .report.appcompat_report import ...`: the new module imports this
one's shared per-change helpers (`_change_to_dict`, `_finding_id`,
`_metadata_dict`, `_fmt_size`, `_VERDICT_EMOJI`), so a static re-export
would close a real import cycle between the two. Resolving on first
attribute access keeps the module-level edge one-way, which is what
`import-cycle-growth` actually checks -- the same shim pattern
`cli_buildsource.py` already uses for its own relocated helpers (root
`AGENTS.md`, "Moving helpers out of a module that re-exports them?").
Both `from abicheck.reporter import appcompat_to_json` and
`reporter.appcompat_to_json` still resolve, which is what the existing
tests use.

Verified: `check_architecture.py` 0 findings, `ruff` and `mypy abicheck/`
clean (573 files), and the 334 appcompat/reporter/confidence tests pass.
